Centralize a `markdown_to_safe_html` function
Summary:
Now that every summary has a `summary_description` field, plugins will
need to serve rendered Markdown to the frontend (assuming that we don't
want to do Markdown rendering in JavaScript, which we don't). This
has security implications (sanitization), so TensorBoard should expose a
function that does the right thing.

An alternative would be to provide a route that provides rendered
Markdown for the `summary_description` only. This has disadvantages:
(1) it would require at least one extra network roundtrip (plugins must
first request the list of tags, then request the rendered description),
(2) it would induce a FoUC-like state where the tags have been fetched
but their metadata has not; and (3) it would not generalize to plugins
that want to safely render other kinds of Markdown, like the text
plugin.

Test Plan:
Run unit tests, and verify that the text and histogram plugins continue
to work.

wchargin-branch: markdown-to-safe-html