Introduce `data_compat`, legacy summary converter
Summary:
As we move from using old-style summaries to using only tensor
summaries, we need to continue supporting the old formats. The only
plugin that has been converted so far is the histogram plugin, and that
plugin manages the compatibility itself. But it would be nicer to have a
unified Event Translation Layer where we convert all the old-style
events to new-style events. Once we do this, plugin code will only have
to deal with new-style events---which is especially attractive because
it makes the plugins more copy-pastable.

This commit introduces `data_compat`, a module that will perform such
conversions. This module will intercept events before the event
accumulator processes them. This is forward-compatible with SQL mode:
we can simply call out to this module from the SQL loader instead of the
event accumulator. For now, `data_compat` converts histogram values and
leaves everything else alone.

I haven't connected this layer to the event accumulator because it has
breaking behavior for the distributions plugin, which depends on
special-casing behavior in the accumulator. The plan to deal with that
is to move the histogram compression logic out of the event accumulator
and into the distributions plugin itself (so that compression is
performed on the fly instead of being precomputed).

Test Plan:
Run the included unit tests, which are comprehensive.

Also, apply the following patch:
```diff
diff --git a/tensorboard/backend/event_processing/BUILD b/tensorboard/backend/event_processing/BUILD
index 3ae126f..4043ca2 100644
--- a/tensorboard/backend/event_processing/BUILD
+++ b/tensorboard/backend/event_processing/BUILD
@@ -82,6 +82,7 @@ py_library(
         ":event_file_loader",
         ":plugin_asset_util",
         ":reservoir",
+        "//tensorboard:data_compat",
         "//tensorboard:expect_tensorflow_installed",
         "//tensorboard/plugins/distribution:compressor",
     ],
diff --git a/tensorboard/backend/event_processing/event_accumulator.py b/tensorboard/backend/event_processing/event_accumulator.py
index 65acc30..c588579 100644
--- a/tensorboard/backend/event_processing/event_accumulator.py
+++ b/tensorboard/backend/event_processing/event_accumulator.py
@@ -23,6 +23,7 @@ import threading

 import tensorflow as tf

+from tensorboard import data_compat
 from tensorboard.backend.event_processing import directory_watcher
 from tensorboard.backend.event_processing import event_file_loader
 from tensorboard.backend.event_processing import plugin_asset_util
@@ -383,6 +384,8 @@ class EventAccumulator(object):
       self._tagged_metadata[tag] = event.tagged_run_metadata.run_metadata
     elif event.HasField('summary'):
       for value in event.summary.value:
+        value = data_compat.translate_value(value)
+
         if value.HasField('metadata'):
           tag = value.tag
           # We only store the first instance of the metadata. This check
```
Verify that the resulting TensorBoard instance has working histograms
from both old- and new-style sources. (You can tell that the old-style
histograms have been converted because the reservoir sampling will be
more aggressive.) Note also that the distributions plugin is marked as
inactive, and that the event accumulator tests fail.

wchargin-branch: data-compat